+++ e2e/test-subjects/remote-react-1/src/RemoteReact1.tsx
@@ -1,10 +1,14 @@
+import { useRuntime } from "@leanjs/react";
 import React from "react";
 import { Link } from "react-router-dom";
 
 export function RemoteReact1() {
+  const runtime = useRuntime();
   return (
     <>
-      <h1>React micro-app 1</h1>
+      <h1>
+        {runtime.context.appName} @ {runtime.context.version}
+      </h1>
       <Link to="/">Home page</Link>
     </>
   );