+++ e2e/tests/integration/remote-react/success.spec.js
@@ -4,7 +4,7 @@ describe("Remote React app: success", () => {
   it("displays a dev remote React app", () => {
     cy.visit("http://localhost:44443");
 
-    cy.contains("React micro-app 1");
+    cy.contains("SelfHosted ( RemoteReact1 ) @ ( 1.6.16 )");
   });
 
   it("displays a prod remote React app", () => {