+++ e2e/tests/integration/shell-next/success.spec.js
@@ -4,6 +4,8 @@ describe("Nextjs shell: success", () => {
   it("displays a remote React app", () => {
     cy.visit("http://localhost:44440");
     cy.contains("h1", "Nextjs Host").should("be.visible");
-    cy.contains("h1", "React micro-app 1").should("be.visible");
+    cy.contains("h1", "NextShell ( RemoteReact1 ) @ 1.0.0").should(
+      "be.visible"
+    );
   });
 });