+++ packages/nuxt/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.2](https://github.com/leanjs/leanjs/compare/@leanjs/nuxt@0.5.1...@leanjs/nuxt@0.5.2) (2022-11-08)
+
+**Note:** Version bump only for package @leanjs/nuxt
+
 ## [0.5.1](https://github.com/leanjs/leanjs/compare/@leanjs/nuxt@0.5.0...@leanjs/nuxt@0.5.1) (2022-11-08)
 
 **Note:** Version bump only for package @leanjs/nuxt