+++ packages/react/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.3](https://github.com/leanjs/leanjs/compare/@leanjs/react@0.17.2...@leanjs/react@0.17.3) (2022-11-08)
+
+**Note:** Version bump only for package @leanjs/react
+
 ## [0.17.2](https://github.com/leanjs/leanjs/compare/@leanjs/react@0.17.1...@leanjs/react@0.17.2) (2022-11-08)
 
 ### Bug Fixes