+++ packages/vue-router/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.2](https://github.com/leanjs/leanjs/compare/@leanjs/vue-router@0.7.1...@leanjs/vue-router@0.7.2) (2022-11-08)
+
+**Note:** Version bump only for package @leanjs/vue-router
+
 ## [0.7.1](https://github.com/leanjs/leanjs/compare/@leanjs/vue-router@0.7.0...@leanjs/vue-router@0.7.1) (2022-11-08)
 
 **Note:** Version bump only for package @leanjs/vue-router